Refactors examples for model specs
