+++ spec/fortnox/api/models/examples/document_base.rb
