+++ spec/fortnox/api/models/invoice_row_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 require 'fortnox/api/models/invoice_row'
-require 'fortnox/api/models/examples/document_row_examples'
+require 'fortnox/api/models/examples/document_row'
 
 RSpec.describe Fortnox::API::Model::InvoiceRow, type: :model do
   subject{ described_class }
