+++ spec/fortnox/api/models/invoice_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 require 'fortnox/api/models/invoice'
 require 'fortnox/api/models/context'
-require 'fortnox/api/models/document_base_examples'
+require 'fortnox/api/models/examples/document_base'
 
 describe Fortnox::API::Model::Invoice, type: :model do
   valid_hash = { customer_number: '12345' }
