+++ spec/fortnox/api/models/order_row_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 require 'fortnox/api/models/order_row'
-require 'fortnox/api/models/examples/document_row_examples'
+require 'fortnox/api/models/examples/document_row'
 
 RSpec.describe Fortnox::API::Model::OrderRow, type: :model do
   valid_hash = { order_quantity: 10.5 }
